+++ public/service/resources.go
@@ -9,6 +9,7 @@ import (
 	"github.com/benthosdev/benthos/v4/internal/bundle"
 	"github.com/benthosdev/benthos/v4/internal/component/cache"
 	"github.com/benthosdev/benthos/v4/internal/component/input"
+	"github.com/benthosdev/benthos/v4/internal/component/output"
 	"github.com/benthosdev/benthos/v4/internal/component/ratelimit"
 	"github.com/benthosdev/benthos/v4/internal/manager/mock"
 )
@@ -96,6 +97,35 @@ func (r *Resources) HasCache(name string) bool {
 	return r.mgr.ProbeCache(name)
 }
 
+// AccessInput attempts to access a input resource by name.
+func (r *Resources) AccessInput(ctx context.Context, name string, fn func(i *ResourceInput)) error {
+	return r.mgr.AccessInput(ctx, name, func(in input.Streamed) {
+		fn(newResourceInput(in))
+	})
+}
+
+// HasInput confirms whether an input with a given name has been registered as a
+// resource. This method is useful during component initialisation as it is
+// defensive against ordering.
+func (r *Resources) HasInput(name string) bool {
+	return r.mgr.ProbeInput(name)
+}
+
+// AccessOutput attempts to access an output resource by name. This action can
+// block if CRUD operations are being actively performed on the resource.
+func (r *Resources) AccessOutput(ctx context.Context, name string, fn func(o *ResourceOutput)) error {
+	return r.mgr.AccessOutput(ctx, name, func(o output.Sync) {
+		fn(newResourceOutput(o))
+	})
+}
+
+// HasOutput confirms whether an output with a given name has been registered as
+// a resource. This method is useful during component initialisation as it is
+// defensive against ordering.
+func (r *Resources) HasOutput(name string) bool {
+	return r.mgr.ProbeOutput(name)
+}
+
 // AccessRateLimit attempts to access a rate limit resource by name. This action
 // can block if CRUD operations are being actively performed on the resource.
 func (r *Resources) AccessRateLimit(ctx context.Context, name string, fn func(r RateLimit)) error {
@@ -109,11 +139,4 @@ func (r *Resources) AccessRateLimit(ctx context.Context, name string, fn func(r 
 // initialisation as it is defensive against ordering.
 func (r *Resources) HasRateLimit(name string) bool {
 	return r.mgr.ProbeRateLimit(name)
-}
-
-// AccessInput attempts to access a input resource by name.
-func (r *Resources) AccessInput(ctx context.Context, name string, fn func(in OwnedInput)) error {
-	return r.mgr.AccessInput(ctx, name, func(in input.Streamed) {
-		fn(OwnedInput{i: in})
-	})
 }