+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - New `cached` processor.
 - Go API: New APIs for registering both metrics exporters and open telemetry tracer plugins.
 - Go API: The stream builder API now supports configuring a tracer, and tracer configuration is now isolated to the stream being executed.
+- Go API: Plugin components can now access input and output resources.
 
 ### Fixed